requirements.txt
pandas
psycopg2-binary
python-dotenv
faker